Refactor the Outbound transport and tls in the configuration file
