+++ V2rayNG/app/src/main/java/com/neko/v2ray/dto/V2rayConfig.kt
@@ -349,139 +349,6 @@ data class V2rayConfig(
                 )
             }
 
-            fun populateTransportSettings(
-                transport: String,
-                headerType: String?,
-                host: String?,
-                path: String?,
-                seed: String?,
-                quicSecurity: String?,
-                key: String?,
-                mode: String?,
-                serviceName: String?,
-                authority: String?
-            ): String? {
-                var sni: String? = null
-                network = if (transport.isEmpty()) NetworkType.TCP.type else transport
-                when (network) {
-                    NetworkType.TCP.type -> {
-                        val tcpSetting = TcpSettingsBean()
-                        if (headerType == AppConfig.HEADER_TYPE_HTTP) {
-                            tcpSetting.header.type = AppConfig.HEADER_TYPE_HTTP
-                            if (!TextUtils.isEmpty(host) || !TextUtils.isEmpty(path)) {
-                                val requestObj = TcpSettingsBean.HeaderBean.RequestBean()
-                                requestObj.headers.Host = host.orEmpty().split(",").map { it.trim() }.filter { it.isNotEmpty() }
-                                requestObj.path = path.orEmpty().split(",").map { it.trim() }.filter { it.isNotEmpty() }
-                                tcpSetting.header.request = requestObj
-                                sni = requestObj.headers.Host?.getOrNull(0)
-                            }
-                        } else {
-                            tcpSetting.header.type = "none"
-                            sni = host
-                        }
-                        tcpSettings = tcpSetting
-                    }
-
-                    NetworkType.KCP.type -> {
-                        val kcpsetting = KcpSettingsBean()
-                        kcpsetting.header.type = headerType ?: "none"
-                        if (seed.isNullOrEmpty()) {
-                            kcpsetting.seed = null
-                        } else {
-                            kcpsetting.seed = seed
-                        }
-                        if (host.isNullOrEmpty()) {
-                            kcpsetting.header.domain = null
-                        } else {
-                            kcpsetting.header.domain = host
-                        }
-                        kcpSettings = kcpsetting
-                    }
-
-                    NetworkType.WS.type -> {
-                        val wssetting = WsSettingsBean()
-                        wssetting.headers.Host = host.orEmpty()
-                        sni = host
-                        wssetting.path = path ?: "/"
-                        wsSettings = wssetting
-                    }
-
-                    NetworkType.HTTP_UPGRADE.type -> {
-                        val httpupgradeSetting = HttpupgradeSettingsBean()
-                        httpupgradeSetting.host = host.orEmpty()
-                        sni = host
-                        httpupgradeSetting.path = path ?: "/"
-                        httpupgradeSettings = httpupgradeSetting
-                    }
-
-                    NetworkType.XHTTP.type -> {
-                        val xhttpSetting = XhttpSettingsBean()
-                        xhttpSetting.host = host.orEmpty()
-                        sni = host
-                        xhttpSetting.path = path ?: "/"
-                        xhttpSettings = xhttpSetting
-                    }
-
-                    NetworkType.H2.type, NetworkType.HTTP.type -> {
-                        network = NetworkType.H2.type
-                        val h2Setting = HttpSettingsBean()
-                        h2Setting.host = host.orEmpty().split(",").map { it.trim() }.filter { it.isNotEmpty() }
-                        sni = h2Setting.host.getOrNull(0)
-                        h2Setting.path = path ?: "/"
-                        httpSettings = h2Setting
-                    }
-
-//                    "quic" -> {
-//                        val quicsetting = QuicSettingBean()
-//                        quicsetting.security = quicSecurity ?: "none"
-//                        quicsetting.key = key.orEmpty()
-//                        quicsetting.header.type = headerType ?: "none"
-//                        quicSettings = quicsetting
-//                    }
-
-                    NetworkType.GRPC.type -> {
-                        val grpcSetting = GrpcSettingsBean()
-                        grpcSetting.multiMode = mode == "multi"
-                        grpcSetting.serviceName = serviceName.orEmpty()
-                        grpcSetting.authority = authority.orEmpty()
-                        grpcSetting.idle_timeout = 60
-                        grpcSetting.health_check_timeout = 20
-                        sni = authority
-                        grpcSettings = grpcSetting
-                    }
-                }
-                return sni
-            }
-
-            fun populateTlsSettings(
-                streamSecurity: String,
-                allowInsecure: Boolean,
-                sni: String?,
-                fingerprint: String?,
-                alpns: String?,
-                publicKey: String?,
-                shortId: String?,
-                spiderX: String?
-            ) {
-                security = if (streamSecurity.isEmpty()) null else streamSecurity
-                if (security == null) return
-                val tlsSetting = TlsSettingsBean(
-                    allowInsecure = allowInsecure,
-                    serverName = if (sni.isNullOrEmpty()) null else sni,
-                    fingerprint = if (fingerprint.isNullOrEmpty()) null else fingerprint,
-                    alpn = if (alpns.isNullOrEmpty()) null else alpns.split(",").map { it.trim() }.filter { it.isNotEmpty() },
-                    publicKey = if (publicKey.isNullOrEmpty()) null else publicKey,
-                    shortId = if (shortId.isNullOrEmpty()) null else shortId,
-                    spiderX = if (spiderX.isNullOrEmpty()) null else spiderX,
-                )
-                if (security == AppConfig.TLS) {
-                    tlsSettings = tlsSetting
-                    realitySettings = null
-                } else if (security == AppConfig.REALITY) {
-                    tlsSettings = null
-                    realitySettings = tlsSetting
-                }
-            }
         }
 
         data class MuxBean(
@@ -723,15 +590,4 @@ data class V2rayConfig(
         return null
     }
 
-    fun toPrettyPrinting(): String {
-        return GsonBuilder()
-            .setPrettyPrinting()
-            .disableHtmlEscaping()
-            .registerTypeAdapter( // custom serialiser is needed here since JSON by default parse number as Double, core will fail to start
-                object : TypeToken<Double>() {}.type,
-                JsonSerializer { src: Double?, _: Type?, _: JsonSerializationContext? -> JsonPrimitive(src?.toInt()) }
-            )
-            .create()
-            .toJson(this)
-    }
 }
