+++ V2rayNG/app/src/main/java/com/neko/v2ray/fmt/FmtBase.kt
@@ -1,11 +1,22 @@
 package com.neko.v2ray.fmt
 
+import android.text.TextUtils
 import com.neko.v2ray.AppConfig
 import com.neko.v2ray.dto.NetworkType
 import com.neko.v2ray.dto.ProfileItem
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.GrpcSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.HttpSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.HttpupgradeSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.KcpSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.TcpSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.TlsSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.WsSettingsBean
+import com.neko.v2ray.dto.V2rayConfig.OutboundBean.StreamSettingsBean.XhttpSettingsBean
 import com.neko.v2ray.extension.isNotNullEmpty
 import com.neko.v2ray.handler.MmkvManager
 import com.neko.v2ray.util.HttpUtil
+import com.neko.v2ray.util.JsonUtil
 import com.neko.v2ray.util.Utils
 import java.net.URI
 
@@ -149,6 +160,146 @@ open class FmtBase {
         }
 
         return dicQuery
+    }
+
+    fun populateTransportSettings(
+        streamSettings: StreamSettingsBean,
+        transport: String,
+        headerType: String?,
+        host: String?,
+        path: String?,
+        seed: String?,
+        quicSecurity: String?,
+        key: String?,
+        mode: String?,
+        serviceName: String?,
+        authority: String?,
+        xhttpMode: String?,
+        xhttpExtra: String?,
+    ): String? {
+        var sni: String? = null
+        streamSettings.network = if (transport.isEmpty()) NetworkType.TCP.type else transport
+        when (streamSettings.network) {
+            NetworkType.TCP.type -> {
+                val tcpSetting = TcpSettingsBean()
+                if (headerType == AppConfig.HEADER_TYPE_HTTP) {
+                    tcpSetting.header.type = AppConfig.HEADER_TYPE_HTTP
+                    if (!TextUtils.isEmpty(host) || !TextUtils.isEmpty(path)) {
+                        val requestObj = TcpSettingsBean.HeaderBean.RequestBean()
+                        requestObj.headers.Host = host.orEmpty().split(",").map { it.trim() }.filter { it.isNotEmpty() }
+                        requestObj.path = path.orEmpty().split(",").map { it.trim() }.filter { it.isNotEmpty() }
+                        tcpSetting.header.request = requestObj
+                        sni = requestObj.headers.Host?.getOrNull(0)
+                    }
+                } else {
+                    tcpSetting.header.type = "none"
+                    sni = host
+                }
+                streamSettings.tcpSettings = tcpSetting
+            }
+
+            NetworkType.KCP.type -> {
+                val kcpsetting = KcpSettingsBean()
+                kcpsetting.header.type = headerType ?: "none"
+                if (seed.isNullOrEmpty()) {
+                    kcpsetting.seed = null
+                } else {
+                    kcpsetting.seed = seed
+                }
+                if (host.isNullOrEmpty()) {
+                    kcpsetting.header.domain = null
+                } else {
+                    kcpsetting.header.domain = host
+                }
+                streamSettings.kcpSettings = kcpsetting
+            }
+
+            NetworkType.WS.type -> {
+                val wssetting = WsSettingsBean()
+                wssetting.headers.Host = host.orEmpty()
+                sni = host
+                wssetting.path = path ?: "/"
+                streamSettings.wsSettings = wssetting
+            }
+
+            NetworkType.HTTP_UPGRADE.type -> {
+                val httpupgradeSetting = HttpupgradeSettingsBean()
+                httpupgradeSetting.host = host.orEmpty()
+                sni = host
+                httpupgradeSetting.path = path ?: "/"
+                streamSettings.httpupgradeSettings = httpupgradeSetting
+            }
+
+            NetworkType.XHTTP.type -> {
+                val xhttpSetting = XhttpSettingsBean()
+                xhttpSetting.host = host.orEmpty()
+                sni = host
+                xhttpSetting.path = path ?: "/"
+                xhttpSetting.mode = xhttpMode
+                xhttpSetting.extra = JsonUtil.parseString(xhttpExtra)
+                streamSettings.xhttpSettings = xhttpSetting
+            }
+
+            NetworkType.H2.type, NetworkType.HTTP.type -> {
+                streamSettings.network = NetworkType.H2.type
+                val h2Setting = HttpSettingsBean()
+                h2Setting.host = host.orEmpty().split(",").map { it.trim() }.filter { it.isNotEmpty() }
+                sni = h2Setting.host.getOrNull(0)
+                h2Setting.path = path ?: "/"
+                streamSettings.httpSettings = h2Setting
+            }
+
+//                    "quic" -> {
+//                        val quicsetting = QuicSettingBean()
+//                        quicsetting.security = quicSecurity ?: "none"
+//                        quicsetting.key = key.orEmpty()
+//                        quicsetting.header.type = headerType ?: "none"
+//                        quicSettings = quicsetting
+//                    }
+
+            NetworkType.GRPC.type -> {
+                val grpcSetting = GrpcSettingsBean()
+                grpcSetting.multiMode = mode == "multi"
+                grpcSetting.serviceName = serviceName.orEmpty()
+                grpcSetting.authority = authority.orEmpty()
+                grpcSetting.idle_timeout = 60
+                grpcSetting.health_check_timeout = 20
+                sni = authority
+                streamSettings.grpcSettings = grpcSetting
+            }
+        }
+        return sni
+    }
+
+    fun populateTlsSettings(
+        streamSettings: StreamSettingsBean,
+        streamSecurity: String,
+        allowInsecure: Boolean,
+        sni: String?,
+        fingerprint: String?,
+        alpns: String?,
+        publicKey: String?,
+        shortId: String?,
+        spiderX: String?
+    ) {
+        streamSettings.security = if (streamSecurity.isEmpty()) null else streamSecurity
+        if (streamSettings.security == null) return
+        val tlsSetting = TlsSettingsBean(
+            allowInsecure = allowInsecure,
+            serverName = if (sni.isNullOrEmpty()) null else sni,
+            fingerprint = if (fingerprint.isNullOrEmpty()) null else fingerprint,
+            alpn = if (alpns.isNullOrEmpty()) null else alpns.split(",").map { it.trim() }.filter { it.isNotEmpty() },
+            publicKey = if (publicKey.isNullOrEmpty()) null else publicKey,
+            shortId = if (shortId.isNullOrEmpty()) null else shortId,
+            spiderX = if (spiderX.isNullOrEmpty()) null else spiderX,
+        )
+        if (streamSettings.security == AppConfig.TLS) {
+            streamSettings.tlsSettings = tlsSetting
+            streamSettings.realitySettings = null
+        } else if (streamSettings.security == AppConfig.REALITY) {
+            streamSettings.tlsSettings = null
+            streamSettings.realitySettings = tlsSetting
+        }
     }
 
     fun resolveHostToIP(server: String?): String {
