+++ V2rayNG/app/src/main/java/com/neko/v2ray/fmt/ShadowsocksFmt.kt
@@ -140,29 +140,37 @@ object ShadowsocksFmt : FmtBase() {
             server.method = profileItem.method
         }
 
-        val sni = outboundBean?.streamSettings?.populateTransportSettings(
-            profileItem.network.orEmpty(),
-            profileItem.headerType,
-            profileItem.host,
-            profileItem.path,
-            profileItem.seed,
-            profileItem.quicSecurity,
-            profileItem.quicKey,
-            profileItem.mode,
-            profileItem.serviceName,
-            profileItem.authority,
-        )
+        val sni = outboundBean?.streamSettings?.let {
+            populateTransportSettings(
+                it,
+                profileItem.network.orEmpty(),
+                profileItem.headerType,
+                profileItem.host,
+                profileItem.path,
+                profileItem.seed,
+                profileItem.quicSecurity,
+                profileItem.quicKey,
+                profileItem.mode,
+                profileItem.serviceName,
+                profileItem.authority,
+                profileItem.xhttpMode,
+                profileItem.xhttpExtra
+            )
+        }
 
-        outboundBean?.streamSettings?.populateTlsSettings(
-            profileItem.security.orEmpty(),
-            profileItem.insecure == true,
-            if (profileItem.sni.isNullOrEmpty()) sni else profileItem.sni,
-            profileItem.fingerPrint,
-            profileItem.alpn,
-            profileItem.publicKey,
-            profileItem.shortId,
-            profileItem.spiderX,
-        )
+        outboundBean?.streamSettings?.let {
+            populateTlsSettings(
+                it,
+                profileItem.security.orEmpty(),
+                profileItem.insecure == true,
+                if (profileItem.sni.isNullOrEmpty()) sni else profileItem.sni,
+                profileItem.fingerPrint,
+                profileItem.alpn,
+                profileItem.publicKey,
+                profileItem.shortId,
+                profileItem.spiderX,
+            )
+        }
 
         return outboundBean
     }
