+++ V2rayNG/app/src/main/java/com/neko/v2ray/fmt/VmessFmt.kt
@@ -177,29 +177,37 @@ object VmessFmt : FmtBase() {
             vnext.users[0].security = profileItem.method
         }
 
-        val sni = outboundBean?.streamSettings?.populateTransportSettings(
-            profileItem.network.orEmpty(),
-            profileItem.headerType,
-            profileItem.host,
-            profileItem.path,
-            profileItem.seed,
-            profileItem.quicSecurity,
-            profileItem.quicKey,
-            profileItem.mode,
-            profileItem.serviceName,
-            profileItem.authority,
-        )
+        val sni = outboundBean?.streamSettings?.let {
+            populateTransportSettings(
+                it,
+                profileItem.network.orEmpty(),
+                profileItem.headerType,
+                profileItem.host,
+                profileItem.path,
+                profileItem.seed,
+                profileItem.quicSecurity,
+                profileItem.quicKey,
+                profileItem.mode,
+                profileItem.serviceName,
+                profileItem.authority,
+                profileItem.xhttpMode,
+                profileItem.xhttpExtra
+            )
+        }
 
-        outboundBean?.streamSettings?.populateTlsSettings(
-            profileItem.security.orEmpty(),
-            profileItem.insecure == true,
-            if (profileItem.sni.isNullOrEmpty()) sni else profileItem.sni,
-            profileItem.fingerPrint,
-            profileItem.alpn,
-            null,
-            null,
-            null
-        )
+        outboundBean?.streamSettings?.let {
+            populateTlsSettings(
+                it,
+                profileItem.security.orEmpty(),
+                profileItem.insecure == true,
+                if (profileItem.sni.isNullOrEmpty()) sni else profileItem.sni,
+                profileItem.fingerPrint,
+                profileItem.alpn,
+                null,
+                null,
+                null
+            )
+        }
 
         return outboundBean
     }
